+++ c/tskit/tables.h
@@ -728,6 +728,7 @@ typedef struct {
 /* This shares an interface with table collection init.
    TODO: review as part of #1720 */
 #define TSK_LOAD_SKIP_TABLES (1 << 1)
+#define TSK_LOAD_SKIP_REFERENCE_SEQUENCE (1 << 2)
 
 /* Flags for table init. */
 #define TSK_NO_METADATA (1 << 0)
@@ -742,6 +743,7 @@ typedef struct {
 #define TSK_CMP_IGNORE_METADATA (1 << 2)
 #define TSK_CMP_IGNORE_TIMESTAMPS (1 << 3)
 #define TSK_CMP_IGNORE_TABLES (1 << 4)
+#define TSK_CMP_IGNORE_REFERENCE_SEQUENCE (1 << 5)
 
 /* Flags for table collection clear */
 #define TSK_CLEAR_METADATA_SCHEMAS (1 << 0)
@@ -3421,6 +3423,8 @@ TSK_CMP_IGNORE_TIMESTAMPS
 TSK_CMP_IGNORE_TABLES
     Do not include any tables in the comparison, thus comparing only the
     top-level information of the table collections being compared.
+TSK_CMP_IGNORE_REFERENCE_SEQUENCE
+    Do not include the reference sequence in the comparison.
 @endrst
 
 @param self A pointer to a tsk_table_collection_t object.
@@ -3479,8 +3483,11 @@ If the file contains multiple table collections, this function will load
 the first. Please see the :c:func:`tsk_table_collection_loadf` for details
 on how to sequentially load table collections from a stream.
 
-If the TSK_LOAD_SKIP_TABLES option is set, only the top-level
-information of the table collection will be read, leaving all tables empty.
+If the TSK_LOAD_SKIP_TABLES option is set, only the non-table information from
+the table collection will be read, leaving all tables with zero rows and no
+metadata or schema.
+If the TSK_LOAD_SKIP_REFERENCE_SEQUENCE option is set, the table collection is
+read without loading the reference sequence.
 
 **Options**
 
@@ -3491,6 +3498,8 @@ TSK_NO_INIT
     Do not initialise this :c:type:`tsk_table_collection_t` before loading.
 TSK_LOAD_SKIP_TABLES
     Skip reading tables, and only load top-level information.
+TSK_LOAD_SKIP_REFERENCE_SEQUENCE
+    Do not load reference sequence.
 
 **Examples**
 
@@ -3539,10 +3548,14 @@ different error conditions. Please see the
 sequentially load tree sequences from a stream.
 
 Please note that this streaming behaviour is not supported if the
-TSK_LOAD_SKIP_TABLES option is set. With this option, only the top-level
-information of the table collection will be read, leaving all tables empty. When
-attempting to read from a stream with multiple table collection definitions and
-the TSK_LOAD_SKIP_TABLES option set, only the top-level information of the first
+TSK_LOAD_SKIP_TABLES or TSK_LOAD_SKIP_REFERENCE_SEQUENCE option is set.
+If the TSK_LOAD_SKIP_TABLES option is set, only the non-table information from
+the table collection will be read, leaving all tables with zero rows and no
+metadata or schema.
+If the TSK_LOAD_SKIP_REFERENCE_SEQUENCE option is set, the table collection is
+read without loading the reference sequence.
+When attempting to read from a stream with multiple table collection definitions
+and either of these two options set, the requested information from the first
 table collection will be read on the first call to
 :c:func:`tsk_table_collection_loadf`, with subsequent calls leading to errors.
 
@@ -3555,6 +3568,8 @@ TSK_NO_INIT
     Do not initialise this :c:type:`tsk_table_collection_t` before loading.
 TSK_LOAD_SKIP_TABLES
     Skip reading tables, and only load top-level information.
+TSK_LOAD_SKIP_REFERENCE_SEQUENCE
+    Do not load reference sequence.
 
 @endrst